chore: split into separate sample files
