+++ spanner/spring-data/src/main/java/com/example/spanner/QuickStartSample.java
@@ -71,16 +71,6 @@ public class QuickStartSample implements CommandLineRunner {
     LOG.info("(SpannerRepositorySample): Running queries.");
     spannerRepositorySample.runRepositoryExample();
 
-    LOG.info("(SpannerTemplateSample): Running a query and an update with a timeout.");
-    spannerTemplateSample.runTemplateTimeoutExample(singer);
-
-    /*
-     This call uses queries defined in SingerRepository. The implementations of those queries
-     are generated by Spring Data Cloud Spanner. These queries can also use timeouts.
-    */
-    LOG.info("(SpannerRepositorySample): Running queries with timeouts.");
-    spannerRepositorySample.runRepositoryTimeoutExample();
-    
     /*
      This call drops both the sample Singer and Album tables.
     */
