+++ spanner/spring-data/src/main/java/com/example/spanner/SpannerTemplateSample.java
@@ -18,16 +18,12 @@ package com.example.spanner;
 
 // [START spring_data_spanner_template_sample]
 import com.google.cloud.spanner.KeySet;
-import com.google.cloud.spanner.SpannerOptions;
-import com.google.cloud.spanner.SpannerOptions.SpannerCallContextTimeoutConfigurator;
 import com.google.cloud.spanner.Statement;
 import com.google.cloud.spring.data.spanner.core.SpannerQueryOptions;
 import com.google.cloud.spring.data.spanner.core.SpannerTemplate;
-import io.grpc.Context;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.threeten.bp.Duration;
 
 /**
  * A quick start code for Spring Data Cloud Spanner. It demonstrates how to use SpannerTemplate to
@@ -50,36 +46,6 @@ public class SpannerTemplateSample {
     List<Singer> allSingers = this.spannerTemplate
         .query(Singer.class, Statement.of("SELECT * FROM Singers"),
                 new SpannerQueryOptions().setAllowPartialRead(true));
-  }
-  
-  public void runTemplateTimeoutExample(Singer singer) {
-    // Query all Singers using a custom timeout value for the query.
-    Context context = Context.current().withValue(SpannerOptions.CALL_CONTEXT_CONFIGURATOR_KEY,
-        SpannerCallContextTimeoutConfigurator.create()
-            .withExecuteQueryTimeout(Duration.ofSeconds(10L)));
-    context.run(() -> {
-      List<Singer> allSingers = this.spannerTemplate.query(Singer.class,
-          Statement.of("SELECT * FROM Singers"),
-          new SpannerQueryOptions().setAllowPartialRead(true));
-    });
-
-    // Delete all singers using a DML statement with a timeout.
-    Context.current().withValue(SpannerOptions.CALL_CONTEXT_CONFIGURATOR_KEY,
-        SpannerCallContextTimeoutConfigurator.create()
-            .withExecuteUpdateTimeout(Duration.ofSeconds(10L))).run(() -> {
-              this.spannerTemplate
-                 .executeDmlStatement(Statement.of("delete from Singers where true"));
-            });
-
-    // Insert a new Singer record using a mutation (Commit) with a timeout.
-    // NOTE: SpannerTemplate uses mutations for executing data modifications. Mutations are applied
-    // to the database using the Commit RPC. This example therefore sets a timeout for the Commit
-    // RPC.
-    Context.current().withValue(SpannerOptions.CALL_CONTEXT_CONFIGURATOR_KEY,
-        SpannerCallContextTimeoutConfigurator.create()
-            .withCommitTimeout(Duration.ofSeconds(10L))).run(() -> {
-              this.spannerTemplate.insert(singer);
-            });
   }
 
 }
